requirements.txt
asyncping3==3.1.2
anyio==4.10.0
discord.py>=2.0.0
